Tock 2.0 TRD: Allow new error codes to be added in the future.
This changes the wording of the error code section to be clear about how the error code list may change in the future. It also modifies the wording to allow us to add error codes in future Tock 2.x versions. Last, it restricts the kernel error code range to 1-1023, reducing the memory needed to store a return code to two bytes.