[WIP] Platform: add iter_drivers method
This adds a method to iterate over all registered drivers of a
Platform-trait implementation. It is particularly useful to
communicate global events such as process start, process exit and
others to drivers.

Work in progress: the iter_drivers method is not yet implemented on
every platform in the Tock repository.