Port MAX17205 system call capsule to 2.0
This capsule is _not_ multi-process friendly, so this port is a tad bit
kludgely (in particular, need to use `MapCell` to store the callback,
unlike with other capsules).

I've also left ReturnCode in place for the internal interface since the
client/driver interface can probably generally be improved with the
removal of ReturnCode, but it's a fairly breaking change.