rv32i: Don't sort PMP regions
The RISC-V spec describes TOR like this:

If TOR is selected, the associated address register forms the top of the address range, and the
preceding PMP address register forms the bottom of the address range. If PMP entry i’s A field is
set to TOR, the entry matches any address y such that pmpaddri−1 ≤ y < pmpaddri (irrespective of
the value of pmpcfgi−1 ). If PMP entry 0’s A field is set to TOR, zero is used for the lower bound,
and so it matches any address y < pmpaddr0.

The inital PMP implementation mis-interpreted this and required soring
the regions and specifically made the previous region match AND then
disabled read/write/access.

This PR changes that operation. Now instead we do the following.

For a region i, we set pmpaddr[i-1] as the start address and pmpaddr[i]
as the end address (same as before). We then set pmpcfg[i-1] to be
disabled, so there is no region there. We then set pmpcfg[i] as TOR and
it matches between pmpaddr[i-1] and pmpaddr[i].

The advantage with this new approach is we no longer have to sort the
regions (which can take time). The main advantage of sorting the regions
is that we could share addresses, for example if a start and end address
are the same we can save a PMP regions by re-using those two addresses.

In reality though this doesn't happen for apps (as flash and memory are
split) and we don't support it anyway.

Sorting the regions will also break if we have a locked region in the
middle of the list.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>