boards: make: Add `install` target
Tock has had a long standing convention of two makefile targets for
boards to load a kernel onto a board.

- flash: Uses JTAG to flash the kernel.
- program: Uses UART and a bootloader to flash the kernel.

This is helpful because some boards support both but require different
hardware, and it gives the user some idea of what tools will be needed
to load the kernel.

However, for new users they must figure out which is appropriate if a
board supports both. Often there is one target that experienced users
just know to use, and we should make this more clear to new users. Also,
having different targets on diferent boards complicates writing a common
tutorial guide for multiple platforms.

This patch adds a new target `make install` that just uses the default
for the platform. The expectation is that most users can install the
kernel using `make install`. The other targets are of course there for
users who need something more specific.