Port IPC driver to Tock 2.0 interface
Mostly a fairly straightforward port, however, I had to change the driver interface
in order to handle 2.0's new allow-semantics. Specifically, allow (both
readonly and readwrite) can no longer return a value, which is how IPC
returned a target descriptor for discovery until now. No matter, the
logic is better suited elsewhere anyway.

Now, discovery is split into an allow_readonly (which is nice so that
large package names can be hardcoded in ROM rather than copied to
memory) followed by a command(IPC, 0). allow_readwrite is _only_ used to
share buffers between apps.

To make this cleaner, I also rejiggerd the arguments to command.
Previous, the first argument to command was the target descriptor for a
notify, while the second argument differentiated between a service and
client notify (the third was unused). Now, the first argument is a
command number, differentiating between a discover (0), service notify
(1), and client notify (2), the second is the target descriptor and the
third is still unused.