arch: cortex: fix MPU rouding calculation
When determining if a `brk()` call is valid for a process, the MPU has
to determine if a subregion (1/8th of the entire app memory region) will
cover the new app break address _without_ also including the kernel
reserved grant region. If so, it configures the MPU and approves the new
app_brk.

Before this math was incorrect if the app break happened to fall on one
of the subregion boundaries. The calculation would include an extra
region, and if that extra region happened to include the grant region
then the brk() would fail, even though it should have been valid. This
fixes the calculation to not over-include subregions in that edge case.