host_emulation: Fix sleep to not use systick
According to kernel::SysTick documentation, systick implementations
do not need to operate during sleep.
And kernel does disable systick by end of Kernel::do_process,
what happens right before going to sleep in kernel_loop.
That causes a crash in current implementation of HostChip::sleep.

A proper solution would be to actively wait for wake up by an interrupt
(from a peripheral) but there is no infrastructure in host_emulation
for that, yet.
The solution is a temporary workaround to wake kernel up periodically.

BUG=None
TEST=make -C boards/host_emulation &&
target/x86_64-unknown-linux-gnu/release/host-testing -a
../libtock-rs/target/release/examples/hello_world

Change-Id: I6fcd209a0158d1f7340d1af688012b29760a4b4c