host_emulation: Send and received UART data over socket
A new socket is bound and a listener thread is started in initialize().
It stores output and input streams every time a new client connection is
established (overwriting any previous streams).
The streams are then used in transmit and receive functions.
The streams access by the main and listener threads are guarded by mutexes.

BUG=None
TEST=With other CLs in platform project.

Change-Id: If1a9fb917f086ef7a599f1d212faca069ce8c6a2