host_emulation: non-blocking operations on UART and asynchronous rx/tx callbacks
Do not block in receive_buffer function on waiting for the data, instead
return immediately and call the ReceiveClient callback once the buffer is filled in.
The callback is called from within Chip::service_pending_interrupts.
Handle the blocking socket reads in a worker thread to no block the kernel.

Call TransmitClient callback similarly outside of the transmit function.

BUG=None
TEST=With other CLs in platform project.

Change-Id: I55fa01afe17b42126295588ebbce93f365136439