AppSlice: fix unsoundness with arbitrary pointers in slices
Rust has strict pointer validity requirements, which -- in part --
also apply to accesses of zero length. In Tock 2.0, AppSlices are
allowed to have arbitrary pointers passed in by userspace as long as
their length is set to be 0. However, it is still not automatically
safe & sound to construct Rust slices from these pointers.

This change ensures that if an AppSlice of length 0 is used (i.e. a
Rust slice is created), the pointer is set to the value returned by
NonNull::dangling(), which is guaranteed to meet the alignment & other
requirements for valid pointers posed by Rust, while not requiring (or
ensuring) that the memory which the pointer points to is valid or
allocated. Therefore, this is only safe for zero-length
accesses (e.g. for zero-length slices).

Signed-off-by: Leon Schuermann <leon@is.currently.online>