ReadWriteAppSlice: require mutable borrow for ReadWrite::mut_map_or
Prior to this change, a ReadWriteAppSlice could be used to obtain two
mutable Rust slices to the same memory region. In Rust, at any given
time, one can have either one mutable reference or any number of
immutable references to some object / memory, and as such this is
incorrect.

This changes ReadWriteAppSlice::mut_map_or to take a mutable reference
to `self` and as such enforces these rules when taking a mutable
borrow of the ReadWriteAppSlice itself.

Reported-by: Johnathan Van Why <jrvanwhy@google.com>
Signed-off-by: Leon Schuermann <leon@is.currently.online>