chips/swerv: Initial commit of SweRV
The SweRV Cores will be used in a variety of Western Digital products in the
near future. All three cores have been open sourced in CHIPS Alliance github.
CHIPS Alliance is an organization that develops open source hardware and open
source software design tools. It is a barrier free, collaborative environment
which lowers the cost of RTL development.

This commit adds initial support for the SweRV chips, with the EH1 PIC.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>