sifive: Move the machine timer from rv32i
The RISC-V architecture requires a machine timer, which is exposed via
the mtime/mtimcmp register. The architecture does not specify where the
mtime/mtimecmp registers are memory mapped though.

The current `arch/rv32i/src/machine_timer.rs` is actually just a SiFive
implementation, so let's move it from the arch crate to the SiFive
crate.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>