allow: move allow system calls handling to sched.rs
This moves the handling of allow-style system calls (both read-only
and read-write) allow from process.rs to sched.rs, given that the
scheduler should have the ultimate authority of interaction between
processes and capsules and implement the system call ABI. This also
mirrors how other system calls are currently handled in Tock.

The logic for checking whether an allow request is valid w.r.t. to the
process' memory regions is still implemented in process.rs.

Signed-off-by: Leon Schuermann <leon@is.currently.online>