TO BE REVERTED: pmp: align base address down, not up.
PMP registers must be set to 4 byte aligned addresses, the code tried
to enforce this by aligning the passed in value up. While this
satisfied the requirement, it was cutting of the base of the address
range, causing seemingly unrelated crashes.

The proper solution for this problem is to refuse to set the register
to an unaligned value, return an error and process the error, to make
sure the user is alerted to the issue at source.

This patch is a temp fix until proper error processing is introduced.

BUG=b:183999620
TEST=the addition of sps_programmer code which caused console app
     crash does not cause the crash any more.

Change-Id: I0e7d2e91605f3c4a27fc6563899868c68572c4d2