rv32i: use atomic CSR instructions for rv32i::support::atomic
Switch rv32i::support::atomic to use the CSRRS and CSRRC atomic RISC-V
CSR instructions exposed via the CSR register wrappers.

Previously, the mstatus CSR was first read and then written to clear
the machine mode interrupts. If an interrupt would arrive between
these instructions, it could have changed mstatus, whereas this
function would write back the previously read and modified register
contents. Furthermore, if an interrupt handler would have cleared MIE
between reading and writing the CSR, this function would set MIE again
after execution of the passed in function.

As far as I can tell, this is not a real issue currently, due to the
way interrupt handlers work and the fact that rv32i::support::atomic
is already executed when the CPU is already in machine mode. However,
it's best not to rely on this and use proper RISC-V instructions.

Additionally, this eliminates 7 instructions and one branch on rustc
1.52.0-nightly.

Signed-off-by: Leon Schuermann <leon@is.currently.online>