capsules/virtio: add VirtIO EntropySource device driver
Implement a driver for the VirtIO EntropySource device. This driver is
an easy target to test and develop the VirtIO infrastructure in Tock.

The VirtIORng device driver offers a kernel::hil::rng::Rng-compatible
interface, such that it can be used by other capsules and exposed to
userspace through the RngDriver capsule.

Signed-off-by: Leon Schuermann <leon@is.currently.online>