kernel: sched: check for process.ready()
If a process is stopped after faulting, it ends up in the `Faulted`
state. Then, if that process still has time in its timeslice remaining,
do_process will try to run it again where it hits a fault because it has
tried to run a faulted/terminated process. To avoid this we must check
if a process actually has any work to do before trying to run it
according to its state.

This patch has the side effect that a yield-ed process with nothing to
do will no longer hit the yield -> next_task() -> None check in the
do_process match statement. I believe this does not matter because both
cases result in a break, and there is no way for a process to have work
scheduled for it between the new check and the existing yield case
check.