boards: nano33: use FaultResponse::StopWithDebug
Because if a process fails immediately the USB/CDC stack doesn't have
time to setup, the panic print message doesn't work. This avoids that
issue by having the kernel just stop the process and notify the user via
a debug message that it faulted. Then the user can use process console
to panic the kernel and see the debug message if they wish.