tock-cells/OptionalCell: rename get-method to extract
As an OptionalCell is designed to look like a Cell which can also be
empty, the get method as currently implemented does not match the
expected behavior (as it will return an `Option<T>` instead of a `T`).
Furthermore, as `OptionalCell::set` takes simply a `T` (not wrapped in
an `Option`), `get` should not return an `Option<T>`.

Accordingly, this commit renames the `OptionalCell::get` method to
`extract`, to distinguish it from the `Cell::get` method.

Fixes: a845aa8e27cfe0 ("tock-cells/OptionalCell: add get-method...")
Suggested-by: Brad Campbell <bradjc5@gmail.com>
Signed-off-by: Leon Schuermann <leon@is.currently.online>