tock-cells/OptionalCell: rename get-method to extract
As an OptionalCell is designed to look like a Cell which can also be
empty, the get method as currently implemented does not match the
expected behavior (as it will return an `Option<T>` instead of a `T`).
Furthermore, as `OptionalCell::set` takes simply a `T` (not wrapped in
an `Option`), `get` should not return an `Option<T>`.

Accordingly, this commit renames the `OptionalCell::get` method to
`extract`, to distinguish it from the `Cell::get` method.

BUG=b:173327144
TEST=make build

Fixes: a845aa8e27cfe0 ("tock-cells/OptionalCell: add get-method...")
Suggested-by: Brad Campbell <bradjc5@gmail.com>
Signed-off-by: Leon Schuermann <leon@is.currently.online>
(cherry picked from commit 3136f00a77173708625256011be54ee24778364f)
Signed-off-by: Stephen Barber <smbarber@google.com>
Change-Id: I5f923c689e3fa530bd61d29945ec1b265a7f0521