kernel: introduce debug::panic_print for boards without LEDs
Introduce an alternative function to debug::panic which does not
require an LED type parameter and returns instead of blinking an LED
in an infinite loop.

This is useful for boards which do not feature any GPIO or LEDs, such
as the SweRVolf and LiteX sim boards running inside Verilator, and
other virtualized boards in the future.

Consequently, the SweRVolf and LiteX sim boards have been changed to
use this function instead.

Signed-off-by: Leon Schuermann <leon@is.currently.online>