virtual_alarm: ensure that we do not miss alarms
If we set alarms for very short duration, then it is possible for these
alarms to get set in the past (which mean they have to wait until a roll
over timer event).

Ensure we are always comparing against the same common time in the past
(prev) as our reference point.

This independently fixes the same issue that is reported here:
https://github.com/tock/tock/issues/1651

BUG=b:183711396
TEST=no more hangs on fw_updater during AP boot
TEST=no more "missed" alarms that must wait until after roll over event.

Change-Id: I7f1bfe703db4698bfd382d825d67c463c7cec6dd