NRF5x 15.4 Radio: replace PPI with software logic
Instead of using the Prorgammable Peripheral Interconnect on the NRF52's
ieee802154 radio driver to retry CCA after a timer expires when it
fails, use simple software logic instead. PPI invokes tasks in one
peripheral (the radio) in response to an event in another (the timer),
without requiring an interrupts serviced in software. PPI is probably _better_ but
it wasn't working and I couldn't quite figure out why it wasn't. The
software logic is just emulating what the PPI is supposed to do.

Because this only gets triggered on a CCA failure, it's pretty
cumbersome to debug.

The best thing to do would be to use PPI as well as shortcuts to
simplify the radio logic significantly, as well as likely improve
transition time performance. I'll leave that for another adventure.