capsules/gpio_async: enforce grant/single-process
Current GPIO HIL semantics notify *all* processes of interrupts to any
pin with interrupts configured (and it's left to higher layers to filter
activity based on which pin generated their interrupt). For Async GPIO,
this is awkward to virtualize (or not), as there is no owning process of
a pin for activity interrupts, but there is for configuration result
interrupts. Also, the underlying hardware likely can't handle multiple
concurrent configurations from multiple apps.

For now then, this capsule tracks a configuration while it is in flight
and notifies the correct process that their configuration has succeeded.
In the rare case where two apps attempt concurrent configuration
requests, the later app will receive `EBUSY`. A retry loop should be
sufficient for most apps to handle this rare case.