teensy40/Makefile: don't assume relative libtock-c path for app-tbf
Don't assume a particular relative path to the libtock-c installation
to look for TBFs (compiled application). Instead, when programming a
kernel together with one or more applications, pass the path to the
TBF applications file in the APP variable.

Furthermore, these changes will cause the combined kernel+application
binary to be rebuilt even if the kernel binary does not change,
whereas previously the `app.elf` file was only dependant on the kernel
elf.

Lastly, the filename of all generated files is prefixed with the
platform name, which is important to avoid collisions in the target
directory.

Signed-off-by: Leon Schuermann <leon@is.currently.online>