kernel: move subscribe handling back to kernel::sched
As part of tock/tock#2462, the handling of subscribe system calls did
involve the Process instance in the original design, to give it
control about subscribed Callbacks. However, that design decision,
originally proposed on tock-2.0-dev, was reverted prior to merging it
into master. This PR still had this change as an artifact of that
period. Hence move subscribe system call handling back to
`kernel::sched`.

Signed-off-by: Leon Schuermann <leon@is.currently.online>