arch/cortex-m: avoid clobbering LLVM-internal used reg r9
As part of a toolchain update to Rust nightly 2021-05-11, on cortex-m,
the register r9 can no longer be marked as clobbered as it is used by
LLVM internally. However, r9 needs to be restored to the process'
register state as it is not hardware-stacked. Hence, mark register r12
as clobbered and store r9 there temporarily, restoring after returning
from the process.

Signed-off-by: Leon Schuermann <leon@is.currently.online>