Fix a bug in the SPI controller capsule.
If userspace performed a read/write that was longer than
the static kernel read buffer, it would only read for the
size of the static kernel buffer. The bug was that the end
of the *total* read was truncated by this buffer length;
it should only be that the individual sub-read is truncated
by this value.

This bug probably went unnnoticed because the static kernel
buffer is 1024, so multi-read/write operations never occurred.

The bug fix is to move the cmp::min from calculating where
to copy into the userspace buffer to calculating where to
copy from the kernel buffer.