Introduce cfg to disable process debug info tracking
This is a re-implementation of a (closed) PR to upstream Tock:
https://github.com/tock/tock/pull/1830. That PR was authored by
Alphan Ulusoy (alphan@google.com). This change introduces a
kernel config option to disable tracking of process debug info
such as number of system calls called and number of timeslice
expirations. The primary downside of disabling this option is
that the process console becomes significantly less useful. The
upside is that it reduces code size by 2048 bytes and reduces system
call overhead by about 200 cycles (per the original PR). The config
can be toggled using cargo features.

BUG=None
TEST=make build; for ti50 with and without the feature enabled.

Cq-Depend: chrome-internal:3933878
Change-Id: I01579a11da9bb58afd78ab9183e07f99c185b9a4