Update rust-toolchain to 2021-05-27
This commit deliberately breaks existing code, as it is only updating
the rust-toolchain for a backport of changes to the tock-registers
crate. A feature previously used by this crate (`const_fn`) does not
work on newer Rust nightlies anymore. In the case of tock-registers it
needs to be replaced with `const_fn_trait_bound`. This updates the
Rust version to properly test the backported changes to the
tock-registers v0.6.0 version with these newer Rust versions.

Signed-off-by: Leon Schuermann <leon@is.currently.online>