Add config option for boards to use minimal or debug panics
This change adds config.debug_panics, a config option which
can be used to remove the contents of process::print_full_process()
and process::print_memory_map(). These two functions are responsible
for 2kB of code, and because of how the ProcessType trait is used as
a trait object their implementations, they are included in the binary
even if never called.

BUG=None

TEST=make build; with and without option passed for Ti50.

Change-Id: Ibeac993dc3f599e6bb81c33ae186abd20c252141