More fully flesh out approach where kernel handles subscribe
This commit is still incomplete. First, num_upcalls comes from
create_grant, but needs to be passed in via main. Second, many of the changes
in the other callback swapping PR need to be incorporated, such as
removing Copy bounds and passing the driver_num into create_grant().
Also, many of the changes to IPC need to be ported (and possibly even
more changes to IPC need to be made).

The only capsule ported to the new interface is alarm. All other
capsules will need to be ported once this is more fleshed out.

Improvements to documentation and naming are also needed. A nice way
for capsules to list all upcalls and populate num_upcalls automatically
would also be nice, but might require macros.