kernel/hil/crc: redesign to use a chunked buffer approach
The previous version of the CRC HIL fostered unsound implementations,
in that it relied on asynchronous processing of input data, but only
took a read-only reference with an anonymous lifetime of the
respective data buffer. When this buffer is used for a DMA operation,
it may already have gone out of scope (for instance, when allocating
on the stack). Therefore, this PR redesigns the CRC HIL interface in
accordance with other current HILs and best practices as documented in
the TRD for HIL design.

Signed-off-by: Leon Schuermann <leon@is.currently.online>