tock-register-interface: move LocalRegisterCopy to own module
Move LocalRegisterCopy to its own module and rexport it through the
top-level crate exports. LocalRegisterCopy is distinct from actual
register types, as it only represents a copy of any other register
type, whether implemented in this crate or outside of it. This is in
contrast to an InMemoryRegister, which represents an actual register
stored in memory. This is further emphasized by the fact that
LocalRegisterCopy does not implement the tock-registers interface
traits (e.g. Readable, Writeable) and uses different method signatures
for manipulating the contained values, and implements the Copy trait.

Hence, this moves it to its own module and reexports it with the
top-level crate exports, as it is of importance regardless of whether
the included register types are used.

Signed-off-by: Leon Schuermann <leon@is.currently.online>