kernel/grant: return concise error on schedule_upcall
Return a Result on GrantUpcallTable::schedule_upcall, indicating
whether an error occurred and the precise reason for the error. Users
of this API may utilize this information to, for example, adjust the
rate at which Upcalls are scheduled or reschedule an Upcall which has
been lost due to the process' task queue being full.

This commit just changes the kernel interfaces, it does not address
all of the new unused Result warnings at all of the capsule scheduling
call sites.

Co-authored-by: Leon Schuermann <leon@is.currently.online>