msp432/adc: document safety invariants on &[u8] to &[u16] conversion
To support ADC operations on `&'static mut [u16]` slice references,
while at the same time using DMA over `&'static mut [u8]` slice
references, the MSP432's ADC implementation features conversion
helpers. While these are unsafe and check some of the required
invariants, these checks and the documentation have been insufficient.

In particular, it is not guaranteed that a given `u8` will be
well-aligned to the alignment constraints of a `u16`. Thus, simply
transmuting a `*mut u8` to a `*mut 16` and converting this back to a
slice can cause undefined behavior.

This does not add additional checks on these functions given they are
already unsafe. Furthermore, this issue likely does not affect the
existing codebase, given that only buffers which were `&'static mut
[u16]` slice references get converted to `&'static mut [u8]`, and back
to `&'static mut [u16]` again. Nonetheless, the presence of checks for
some of the required invariants could be interpreted as a function
which is safe to use, even though it is marked unsafe.

Instead, this PR adds a clear indication of the required invariants
and the consequences of a violation of any given requirement.

Signed-off-by: Leon Schuermann <leon@is.currently.online>