trd/104-syscalls: clarify guarantees w.r.t. returned Allow buffers
This clarifies what guarantees the kernel provides w.r.t. the returned
buffer information as part of an Allow system call.

These changes are necessary, because the kernel does not enforce that
drivers always return a zero-length buffer on the first call to Allow,
and return the previously allowed buffer on subsequent calls to Allow.

The wording to use in the TRD104 Syscalls document is tricky, as there
are two perspectives on this issue, which must be clearly separated:

- *the capsule implementation*: a capsule MUST return the correct buffer
  to be correct and compliant to the specification.

- *the system call / kernel behavior*: applications MUST NOT rely on
  previously allowed buffers to be returned, as least w.r.t. to the
  current kernel implementation.

Unfortunately, TRD104 prior to these changes expresses: "Calling a
Read-Write Allow system call returns a buffer (address and length). On
the first call to a Read-Write Allow system call, the kernel returns a
zero-length buffer. Subsequent successful calls to Read-Write Allow
return the previous buffer passed."

This sets expectations for the userspace / kernel ABI, which are not
validated or enforced as part of the current kernel
specification. Thus, this commit changes the wording of TRD104 to
outline these rules for drivers, but also clarify that these are not
enforced as of now, but may be enforced in the future. Userspace
applications must expect that drivers are not behaving as specified
and should employ proper error handling.

Signed-off-by: Leon Schuermann <leon@is.currently.online>