Fix the SPI MUX to store configuration on each client and restore it for each operation. Update the SPI HIL a bit to be more in line with the HIL design TRD (add set_client to the traits).
One major note: the SPI virtual devices have long had an insidious
bit of logic that prevents the callback from being changed: calling
set_client inserted the device into the device list. Therefore if
you called it twice you could insert it twice and create an
infinite list. This commit changes the logic to be in line with
what UART does: there is a `setup` call that inserts the device
into the MUX list, and a separate `set_client` call.

This means that we need to check all code that instantiates a SPI
device and make sure it calls `setup`. I think I got them all.

I tested this with the new `spi_loopback` tests in imix. This starts
up two SPI master devices with different speeds and has them both
write. I checked that they alternate in speeds and the speeds are
correct. So the configuration is being set on each operation.