Add a tool for identifying panics in a Tock binary
This tool takes in an elf file and outputs a list of source
code locations which call functions in the core library that
can panic. It is currently functional, but does not guarantee to
catch all panics, and some reported source locations are only
rough estimates, especially when there are several layers of
inlining.