Fix the multi_alarm_test to provide output, refactor it
This PR fixes the multi_alarm_test so it provides output when
called using the component, and operates silently when run as part
of the OT test suite. It also removes the duplicate code in every board
file for this test, switching over all main.rs files to use the
component instead.