boards/litex: bump targeted tock-litex release
The previous release updated the targeted version of the VexRiscv
CPU. The PMP implementation of VexRiscv experienced a regression in
that the TOR addressing mode is no longer supported. This collides
with the expectations of the current PMP implementation in Tock. For
Tock 2.0, this simply downgrades the VexRiscv CPU to an older version
where TOR addressing is still supported. In the future, either the
VexRiscv CPU will need to support TOR again, or an alternative PMP
implementation for Tock needs to be written.

Signed-off-by: Leon Schuermann <leon@is.currently.online>