Changed length behavior based on discussion in core call on 8/13/21.
It used to be that the transfer length is the minimum of the buffer sizes
and the len parameter. This changes it so if the len parameter is longer
than one of the buffers, the call to read_write_bytes returns an
`Err(SIZE)`. This is consistent with what the UART does.