capsules/alarm: remove command 4 "set absolute expiration"
Command number 4 in the Alarm driver sets an absolute expiration time
directly from userspace. This has been superseded by command number 6,
which sets an absolute expiration time with a reference
timestamp. This prevents wraparound bugs, which can could cause the
alarm to be delayed significantly.

Command number 4 has not been used in libtock-c for some time and has
been marked for deprecation in the Tock source code. This removes it
prior to the Tock 2.0 release.

Signed-off-by: Leon Schuermann <leon@is.currently.online>