boards/litex/sim: add GPIO, button and LED drivers
Instantiates the GPIO controller (gpio0) driver for the LiteX
simulation board. Because the LiteX simulation does not have any real
IO pins, there is no fixed assignment to onboard peripherals. Thus,
the 32 IO pins present in the simulation are split up as follows:

     0 -  7: LEDs
     8 - 15: Buttons
    16 - 31: other GPIOs (assigned indices 16 - 31)

This pins are assigned to the respective userspace drivers as
documented.

Signed-off-by: Leon Schuermann <leon@is.currently.online>