components/button: support both referenced and owned GPIO pins
To support easy instantiaion of a button driver using owned GPIO pin
values, this adds a `button_component_helper_owned!` macro as an
alternative to the `button_component_helper!` macro. This will store
the owned GPIO values in a `static_ref!` internally.

The motivation for this change is further outlined in
8bf3913a03e76c ("components/gpio: support both referenced and owned
GPIO pins").

Signed-off-by: Leon Schuermann <leon@is.currently.online>