components/gpio: support both referenced and owned GPIO pins
Before this change the `gpio_component_helper!` macro could only take
references to GPIO pins. This works fine for most boards, which will
have a fixed number of GPIO pins which are statically allocated within
the respective chip crate.

However, it is also useful for the `gpio_component_helper!` macro to
support passing in owned GPIO pins, which are then internally
allocated through `static_init!`. Specifically, dynamic platfoms such
as LiteX do not have a fixed number of GPIOs on every board or even in
different configurations. Thus it is not trivial to preallocate all
GPIO pins within the board crate and take a reference to
them. Instead, GPIO pin instances are created on the fly and passed to
the caller as owned objects (holding a reference to the GPIO
controller). This can also reduce resource usage, as unused GPIO pins
will not consume a significant amount of memory.

To support easy instantiaion of a GPIO driver using such owned GPIO
pin structs, this adds a second macro called
`gpio_component_helper_owned!` to support passed in GPIO pins as owned
values. For owned values, the `gpio_component_helper_owned!` will
allocate memory to store them internally.

Signed-off-by: Leon Schuermann <leon@is.currently.online>