boards/litex/sim: bump targeted tock-litex release
The upstream VexRiscv CPU has changed its PMP implementation to only
support NAPOT addressing. This makes the prebuilt VexRiscv_Secure
target incompatible with the expectations of Tock. However, by
patching in the old VexRiscv PMP implementation again we can use TOR
addressing again. Along with the switch to the VexRiscv_TockSecureIMC
custom CPU target we can take advantage of the RISC-V IMC support.

Furthermore, this adds a GPIO core to the simulation in anticipation
of the LiteX simulator-based CI workflow.

Signed-off-by: Leon Schuermann <leon@is.currently.online>