boards/Makefile.common: allow specifying system LLVM toolchain
The Makefile contains logic such that, when the default toolchain
parameter of "llvm" is provided, it will attempt to search for an LLVM
toolchain in the Rust sysroot path. This will fail when one wants to
use the system's LLVM toolchain (in PATH as llvm-*).

This commit changes the default to the more meaningful
"llvm-tools-preview" magic value. It clearly indicates that this
refers to the toolchain installed by the Rust "llvm-tools-preview"
component and that this string will cause the infrastructure to
resolve this toolchain. This allows developers to once again override
this parameter to use their system's LLVM toolchain.

Signed-off-by: Leon Schuermann <leon@is.currently.online>